When restarting an image, finer boot process.
closes #43

+++ files/base/boot.sh
@@ -3,5 +3,5 @@
 set -xe
 
 cp /var/templates/base/README.md /var/www/
-cp /var/templates/base/config/nginx.conf /var/www/config/
-cp -r /var/templates/base/public /var/www/
+[ ! -f /var/www/config/nginx.conf ] && cp /var/templates/base/config/nginx.conf /var/www/config/
+[ ! -d /var/www/public ] && cp -r /var/templates/base/public /var/www/
