+++ files/python/boot.sh
@@ -3,10 +3,14 @@
 set -xe
 
 cp /var/templates/python/README.md /var/www/
-cp /var/templates/python/config/nginx.conf /var/www/config/nginx.conf
-cp /var/templates/python/config/uwsgi.ini /var/www/config/uwsgi.ini
-cp -r /var/templates/python/app /var/www/app
+[ ! -f /var/www/config/nginx.conf ] && cp /var/templates/python/config/nginx.conf /var/www/config/nginx.conf
+[ ! -f /var/www/config/uwsgi.ini ] && cp /var/templates/python/config/uwsgi.ini /var/www/config/uwsgi.ini
 
-# configure python
-python3.6 -m venv /var/www/app/venv
-/var/www/app/venv/bin/pip --no-cache-dir install -U pip
+if [ ! -d /var/www/app ]
+then
+    cp -r /var/templates/python/app /var/www/app
+
+    # configure python
+    python3.6 -m venv /var/www/app/venv
+    /var/www/app/venv/bin/pip --no-cache-dir install -U pip
+fi
